feat(account): added error message handling in Skin Upload

+++ app/api/users/route.ts
@@ -1,5 +1,7 @@
 import prisma from "@/app/lib/db";
 import logger from "@/app/utils/logger";
+import { getToken } from "next-auth/jwt";
+import { NextRequest } from "next/server";
 
 export async function GET(request: Request) {
   try {
@@ -21,17 +23,34 @@ export async function DELETE(request: Request) {
   const url = new URL(request.url);
   const searchParams = url.searchParams;
   const userId = searchParams.get('id');
+  const token = await getToken({ req: request as NextRequest, secret: process.env.NEXTAUTH_SECRET });
   
   if (!userId) {
     return new Response(JSON.stringify({ message: "Veuillez saisir un identifiant d'utilisateur" }), { status: 400 });
   }
 
+  if (token && token.id === userId) {
+    return new Response(JSON.stringify({ message: "Impossible d'effectuer cette opération."}), { status: 401 });
+  }
+
   try {
+
+    const currentUser = await prisma.users.findUnique({
+      where: {
+        id: userId
+      }
+    });
+
+    if (!currentUser || currentUser.admin) {
+      return new Response(JSON.stringify({ message: "Impossible d'effectuer cette opération."}), { status: 401 });
+    }
+
     const user = await prisma.users.delete({
       where: {
         id: userId
       }
     });
+
     return new Response(JSON.stringify({ message: 'Success', user: { id: user.id, name: user.name }}), { status: 200 });
   } catch (e) {
     logger.error(e)
